Fix json/xml file handling from the favicon service
The online favicons service is returning valid content-type headers (maybe a recent change, though it's sometime in the last few months), which is causing node-rest-client to convert the raw json buffer into an actual JS object. This then breaks `new Buffer()`, which expects something that `must start with number, buffer, array or string`.

This changes fixes that to re-json and re-xml files based on their extensions. (Otherwise it involvex more drastic changes/fixes to node-rest-client and its horrible API)